bigqueryBTC/scripts/requirements.txt
google-cloud-storage==2.7.0
google-cloud-bigquery==3.5.0
google-cloud-pubsub==2.13.0
google-cloud-scheduler==2.8.0
pandas==2.0.3
yfinance==0.2.18
python-dotenv==1.0.0